Move types to other file, separate late fusion from AdvCP manager
