+++ opencda/core/attack/advcp/3d_models/spoof/car_mesh_divide.pkl
