+++ opencda/core/attack/advcp/__init__.py
@@ -1,6 +1,14 @@
 """AdvCP attack helpers."""
 
+from .attack_helper import AdvCPAttackHelper, AdvCPCarMeshHelper
 from .adv_coperception_model_manager import AdvCoperceptionModelManager
 from .early_fusion_attack import AdvCoperceptionEarlyFusionAttack
+from .late_fusion_attack import AdvCoperceptionLateFusionAttack
 
-__all__ = ["AdvCoperceptionModelManager", "AdvCoperceptionEarlyFusionAttack"]
+__all__ = [
+    "AdvCPAttackHelper",
+    "AdvCPCarMeshHelper",
+    "AdvCoperceptionModelManager",
+    "AdvCoperceptionEarlyFusionAttack",
+    "AdvCoperceptionLateFusionAttack",
+]
